[Refactor] move `nodeHasProperty` implementation to `Utils`
